+++ SlyAPI-FSharp.Tests/TestSerde.fs
@@ -12,5 +12,16 @@ type TestSerde () =
     [<TestMethod>]
     member _.TestOAuth2FromFiles () =
         let auth = OAuth2("test app.json", "test user.json")
+        printfn $"{auth.App}"
+        printfn $"{auth.User}"
 
-        ()
+        Assert.IsNotNull auth.App.Id
+        Assert.IsNotNull auth.App.Secret
+        Assert.IsNotNull auth.App.AuthUrl
+        Assert.IsNotNull auth.App.TokenUrl
+        
+        Assert.IsNotNull auth.User.Expires
+        Assert.IsNotNull auth.User.RefreshToken
+        Assert.IsNotNull auth.User.Scopes
+        Assert.IsNotNull auth.User.Token
+        Assert.IsNotNull auth.User.TokenType