+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## [Unreleased]
 
+## [0.0.5] - 2023-02-07
+
+### Added
+
+- Serialization.SnakeCaseNamingPolicy
+
+### Fixed
+
+- OAuth2 and WebAPI use SnakeCaseNamingPolicy
+
+## [0.0.4] - 2023-02-07
+
+### Fixed
+
+- OAuth2 two file path constructor
+
 ## [0.0.3] - 2023-02-07
 
 ### Added